+++ src/main/java/com/study/controller/cnu/CnuBoardController.java
@@ -1,4 +1,4 @@
-package com.study.controller;
+package com.study.controller.cnu;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
