+++ src/main/java/com/study/controller/cnu/HelloWorldController.java
@@ -1,4 +1,4 @@
-package com.study.controller;
+package com.study.controller.cnu;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
