+++ src/main/java/com/study/domain/sample/Board.java
@@ -1,4 +1,4 @@
-package com.study.domain;
+package com.study.domain.sample;
 
 import java.util.Date;
