+++ src/main/java/com/study/repository/mybatis/CnuRepository.java
@@ -2,7 +2,7 @@ package com.study.repository.mybatis;
 
 import java.util.List;
 
-import com.study.domain.Board;
+import com.study.domain.sample.Board;
 
 public interface CnuRepository {
 	List<Board> selectBoardList();
