+++ src/main/java/com/study/service/cnu/CalculateService.java
@@ -1,4 +1,4 @@
-package com.study.service;
+package com.study.service.cnu;
 
 import java.rmi.UnexpectedException;
